Restructuring Python application
