+++ Website/Documentation/docs/users.rst
@@ -1,0 +1,35 @@
+User Functionality
+===================
+
+JustHealth provides a number of functionalities associated with user management.
+
+The application has a number of different types of users, including:
+
+1. Patients
+2. Carers
+3. Relatives
+4. Medical Professionals
+
+Register
+--------
+
+The registration function is a simple form where a user can input their account information.
+
+It takes the following fields:
+
+1. Username
+2. First Name
+3. Surname
+4. Date of Birth
+5. Gender
+6. Account Type
+7. Date of Birth
+8. Email Address
+9. Password
+10. Confirm Password
+11. Terms and Conditions Accept
+
+These are passed to the :applicaton:func:`registration` function  
+
+.. automodule:: application
+    :members: registration
