+++ Website/justHealthServer/generalViews.py
@@ -1,0 +1,31 @@
+from justHealthServer import *
+
+from functools import wraps
+
+# Decorator Functions
+def needLogin(f):
+  @wraps(f)
+  def loginCheck(*args, **kwargs):
+    try:
+      session['username']
+    except KeyError, e:
+      # session['username'] doesn't exist, kick to login screen
+      return redirect(url_for('login'))
+    # User logged in, continue as normal
+    return f(*args, **kwargs)
+  return loginCheck
+
+@app.route('/')
+@needLogin
+def index():
+    return session['username']
+
+@app.route('/termsandconditions')
+def terms():
+  return render_template('termsandconditions.html')
+
+@app.route('/logout')
+@needLogin
+def logout():
+  session.pop('username', None)
+  return redirect(url_for('index'))
