+++ Website/justHealthServer/static/js/script.js
