+++ Website/runServer.py
@@ -1,0 +1,2 @@
+from justHealthServer import app
+app.run(debug=True, port=9999)
